3d wave fsi also done
